[ADD] #7 SQLite/Room 설정, Dao/Entity 생성, Query/Coroutine작성
Former-commit-id: 00138ed7aa4c1d9602f3f88f9c5e50acb4d367bc

+++ Smile_Android/app/src/main/java/com/ssafy/smile/data/remote/repository/PortfolioRepositoryImpl.kt
@@ -43,14 +43,14 @@ class PortfolioRepositoryImpl(private val portfolioRemoteDataSource: PortfolioRe
             portfolioRemoteDataSource.uploadPost(images)
         }
     }
-    override suspend fun uploadPost(latitude: Float, longitude: Float,
+    override suspend fun uploadPost(latitude: Double, longitude: Double,
                                     detailAddress: String, category: String, images : List<MultipartBody.Part>){
         safeApiCall(_postUploadResponseLiveData){
             portfolioRemoteDataSource.uploadPost(latitude, longitude, detailAddress, category, images)
         }
     }
 
-    override suspend fun uploadPost(latitude: Float, longitude: Float,
+    override suspend fun uploadPost(latitude: Double, longitude: Double,
                                     detailAddress: String, category: String, images : HashMap<String, List<MultipartBody.Part>>){
         safeApiCall(_postUploadResponseLiveData){
             portfolioRemoteDataSource.uploadPost(latitude, longitude, detailAddress, category, images)
